requirements.txt
fastapi
uvicorn
python-telegram-bot
pyswisseph
sqlalchemy
openai
python-dotenv
